feat: migrate package to typescript
