+++ .circleci/config.yml
@@ -1,56 +1,69 @@
 version: 2
 
 docker_defaults: &docker_defaults
-    docker:
-        - image: circleci/node:8.11.1-browsers
-    working_directory: ~/project/repo
+  docker:
+    - image: circleci/node:12.11.1-browsers
+  working_directory: ~/project/repo
 
 attach_workspace: &attach_workspace
-    attach_workspace:
-        at: ~/project
+  attach_workspace:
+    at: ~/project
 
 install_steps: &install_steps
-    steps:
-        - checkout
-        - restore_cache:
-            name: Restore node_modules cache
-            keys:
-                - dependency-cache-{{ .Branch }}-{{ checksum "package.json" }}
-                - dependency-cache-{{ .Branch }}-
-                - dependency-cache-
-        - run:
-            name: Installing Dependencies
-            command: |
-                npm install --silent
-        - save_cache:
-            name: Save node_modules cache
-            key: dependency-cache-{{ .Branch }}-{{ checksum "package.json" }}
-            paths:
-                - node_modules/
-        - persist_to_workspace:
-            root: ~/project
-            paths:
-                - repo
+  steps:
+    - checkout
+    - restore_cache:
+        name: Restore node_modules cache
+        keys:
+          - dependency-cache-{{ .Branch }}-{{ checksum "package.json" }}
+          - dependency-cache-{{ .Branch }}-
+          - dependency-cache-
+    - run:
+        name: Installing Dependencies
+        command: |
+          yarn install --silent --frozen-lockfile
+    - save_cache:
+        name: Save node_modules cache
+        key: dependency-cache-{{ .Branch }}-{{ checksum "yarn.lock" }}
+        paths:
+          - node_modules/
+    - persist_to_workspace:
+        root: ~/project
+        paths:
+          - repo
 
 workflows:
-    version: 2
-    build_pipeline:
-        jobs:
+  version: 2
+  build_pipeline:
+    jobs:
+      - build
+      - unit_test:
+          requires:
             - build
-            - unit_test:
-                requires:
-                    - build
+      - bundle_size:
+          requires:
+            - build
+      # - end_to_end:
+      #     requires:
+      #         - build
 jobs:
-    build:
-        <<: *docker_defaults
-        <<: *install_steps
-    unit_test:
-        <<: *docker_defaults
-        steps:
-            - *attach_workspace
-            - run:
-                name: Running unit tests
-                command: |
-                  sudo npm run editorconfig-tools-check
-                  sudo npm test -- -R dot --check-leaks
-                  sudo npm run coveralls
+  build:
+    <<: *docker_defaults
+    <<: *install_steps
+  unit_test:
+    <<: *docker_defaults
+    steps:
+      - *attach_workspace
+      - run:
+          name: Running unit tests
+          command: |
+            sudo yarn test:ci
+  bundle_size:
+    <<: *docker_defaults
+    steps:
+      - *attach_workspace
+      - run:
+          name: Checking bundle size
+          command: |
+            sudo yarn build
+            sudo yarn bundlesize
