+++ .editorconfig
@@ -1,13 +1,19 @@
+# EditorConfig helps developers define and maintain consistent
+# coding styles between different editors and IDEs
+# http://editorconfig.org
+
 root = true
 
 [*]
+# Change these settings to your own preference
 indent_style = space
 indent_size = 2
+
+# We recommend you to keep these unchanged
 end_of_line = lf
 charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
-max_line_length = 120
 
 [*.md]
 trim_trailing_whitespace = false
