+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,19 @@
+### Minimal example
+
+Add here an online minimal example and share the link here, if the case.
+
+### Expected behaviour
+
+I thought that by going to the page '...' and do the action '...' then '...' would happen.
+
+### Actual behaviour
+
+Instead of '...', what I saw was that '...' happened instead.
+
+### How to simulate the behaviour (if applicable)
+
+This is the example of the current behaviour https://stackblitz.com/edit/hex-to-css-filter-playground
+
+### Browsers/OS/Platforms affected
+
+Please specify the affected browsers/OS/platforms.
