+++ .gitignore
@@ -1,8 +1,15 @@
-.nyc_output/
-node_modules/
-coverage/
-reporter/
+lib
+npm-debug.log
 .DS_Store
-*.swp
-*.swo
-*.log
+node_modules
+coverage
+package
+.nyc_output
+dist
+.jest/
+out-tsc/
+*.js
+*.d.ts
+!jest.*.js
+!scripts/*
+*.tgz
