+++ .nvmrc
@@ -1,1 +1,1 @@
-v8.9.3
+v12.3.1
