+++ package.json
@@ -2,44 +2,121 @@
   "name": "hex-to-css-filter",
   "version": "1.0.3",
   "description": "hex-to-css-filter - Easy way to generate colors from HEX to CSS Filters",
-  "main": "index.js",
-  "types": "index.d.ts",
-  "scripts": {
-    "editorconfig-tools-check": "editorconfig-tools check ./index.js ./test/**",
-    "lint": "eslint .",
-    "coveralls": "cat ./coverage/lcov.info | ./node_modules/.bin/coveralls && rm -rf coverage",
-    "test": "NODE_ENV=test nyc --reporter=lcov --reporter=html mocha --recursive test",
-    "version": "version-changelog CHANGELOG.md && changelog-verify CHANGELOG.md && git add CHANGELOG.md"
-  },
   "author": "Will Mendes <willmendesneto@gmail.com>",
-  "license": "MIT",
-  "directories": {
-    "test": "test"
-  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/willmendesneto/hex-to-css-filter.git"
   },
-  "keywords": [],
   "bugs": {
     "url": "https://github.com/willmendesneto/hex-to-css-filter/issues"
   },
   "homepage": "https://github.com/willmendesneto/hex-to-css-filter#readme",
+  "sideEffects": false,
+  "license": "MIT",
+  "browser": "dist/umd/hex-to-css-filter.js",
+  "jsnext:main": "dist/esm/index.js",
+  "module": "dist/esm/index.js",
+  "main": "dist/cjs/index.js",
+  "es2015": "dist/cjs/index.js",
+  "cjs": "dist/cjs/index.js",
+  "types": "dist/umd/index.d.ts",
+  "keywords": [
+    "hex-to-css-filter",
+    "css",
+    "filter",
+    "hex",
+    "color"
+  ],
+  "dependencies": {
+    "tslib": "^1.10.0"
+  },
   "devDependencies": {
+    "@types/jest": "^24.0.18",
+    "@types/node": "^12.0.10",
+    "@typescript-eslint/eslint-plugin": "^2.3.3",
+    "@typescript-eslint/parser": "^2.3.3",
+    "bundlesize": "^0.18.0",
     "changelog-verify": "^1.1.0",
-    "coveralls": "^3.0.2",
-    "editorconfig-tools": "^0.1.1",
-    "eslint": "^5.0.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.10.0",
-    "eslint-plugin-node": "^7.0.0",
-    "eslint-plugin-promise": "^4.0.0",
-    "eslint-plugin-standard": "^4.0.0",
-    "mocha": "^5.0.5",
-    "nyc": "^12.0.1",
-    "proxyquire": "^2.0.1",
-    "sinon": "^6.0.0",
+    "coveralls": "^3.0.7",
+    "eslint": "^6.5.1",
+    "eslint-config-prettier": "^6.4.0",
+    "eslint-plugin-compat": "^3.3.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "fs-extra": "^8.1.0",
+    "glob": "^7.1.4",
+    "husky": "^3.0.9",
+    "jest": "^24.9.0",
+    "jsdom": "^15.2.0",
+    "lint-staged": "^9.4.2",
+    "prettier": "^1.18.2",
+    "rollup": "^1.26.3",
+    "ts-jest": "^24.1.0",
+    "ts-node": "^8.4.1",
+    "typescript": "^3.4.3",
+    "typings": "^2.1.1",
+    "uglifyjs": "^2.4.11",
     "version-changelog": "^3.1.0"
   },
-  "dependencies": {}
+  "engines": {
+    "node": ">=6.10.2"
+  },
+  "scripts": {
+    "compile": "tsc",
+    "clean": "rm -rf ./dist ./.jest ./coverage ./lib",
+    "build": "yarn build:es2015 && yarn build:cjs && yarn build:esm && yarn build:umd",
+    "build:umd": "rollup dist/esm/index.js --format umd --name PerfMarks -o dist/umd/hex-to-css-filter.js && yarn build:umd:min",
+    "build:umd:min": "uglifyjs --compress --mangle --screw-ie8 --comments -o dist/umd/hex-to-css-filter.min.js -- dist/umd/hex-to-css-filter.js && gzip dist/umd/hex-to-css-filter.min.js -c > dist/umd/hex-to-css-filter.min.js.gz",
+    "build:es2015": "tsc --module es2015 --target es2015 --outDir dist/es2015",
+    "build:esm": "tsc --module esnext --target es5 --outDir dist/esm",
+    "build:cjs": "tsc --module commonjs --target es5 --outDir dist/cjs",
+    "test": "jest",
+    "pretest:ci": "yarn lint",
+    "test:ci": "jest --coverage --coverageReporters=text-lcov | coveralls",
+    "check-coverage": "cat ./coverage/lcov.info | ./node_modules/.bin/coveralls && rm -rf coverage",
+    "bundlesize": "bundlesize",
+    "lint": "eslint '{scripts,src}/**/*.[tj]s'",
+    "lint:fix": "prettier --no-editorconfig --write",
+    "version": "version-changelog CHANGELOG.md && changelog-verify CHANGELOG.md && git add CHANGELOG.md"
+  },
+  "bundlesize": [
+    {
+      "path": "./dist/esm/index.js",
+      "maxSize": "58B"
+    },
+    {
+      "path": "./dist/es2015/index.js",
+      "maxSize": "58B"
+    },
+    {
+      "path": "./dist/cjs/index.js",
+      "maxSize": "155B"
+    },
+    {
+      "path": "./dist/umd/hex-to-css-filter.js",
+      "maxSize": "3.06KB"
+    },
+    {
+      "path": "./dist/umd/hex-to-css-filter.min.js",
+      "maxSize": "2.05KB"
+    }
+  ],
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.{js,ts}": [
+      "prettier --no-editorconfig --write",
+      "git add"
+    ]
+  },
+  "browserslist": [
+    "last 1 chrome versions",
+    "last 1 edge versions",
+    "last 1 firefox versions",
+    "last 1 safari versions",
+    "last 1 and_chr versions",
+    "last 1 ios_saf versions"
+  ]
 }
